+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+<a name="3.0.1"></a>
+ * Fix isUrl check
+ * Add array of xray selector =>  x(url ,'.tag', [x("a", ["@text"])])
+
 <a name="3.0.0"></a>
  * Fork from x-ray 2.3.2 (https://github.com/lapwinglabs/x-ray/)
  * Few changes on the original instantiation api